webpage/config/requirements.txt
backports-abc==0.4
decorator==4.0.4
Django==1.9
django-audit-log==0.7.0
django-autocomplete-light==2.2.10
django-crispy-forms==1.5.2
django-extensions==1.5.9
djangorestframework==3.2.4
ipykernel==4.1.1
ipython==4.0.0
ipython-genutils==0.1.0
ipywidgets==4.1.1
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
MarkupSafe==0.23
mistune==0.7.1
mysqlclient==1.3.6
nbconvert==4.1.0
nbformat==4.0.1
notebook==4.0.6
path.py==8.1.2
pickleshare==0.5
Pygments==2.0.2
pyzmq==15.0.0
qtconsole==4.1.0
requests==2.8.1
simplegeneric==0.8.1
six==1.10.0
tornado==4.3
traitlets==4.0.0
wheel==0.26.0
